// src/adm/ordenesPago/interfaces/impuestoDocumentosOp/deleteImpuestoDocumentosOp.ts
export interface IResponseDeleteImpuestoDocumentosOp {
  data: IDeleteImpuestoDocumentosOp
  isValid: boolean
  linkData: any
  linkDataArlternative: any
  message: string
  page: number
  totalPage: number
  cantidadRegistros: number
  total1: number
  total2: number
  total3: number
  total4: number
}

export interface IDeleteImpuestoDocumentosOp {
  codigoImpuestoDocumentoOp: number;
}